requirement.txt
Pillow==9.4.0
Telethon==1.26.1
beautifulsoup4==4.11.1
requests==2.28.2
python-dotenv==0.21.0